fibonacci_monitors/requirements.txt
requests>=2.31.0
matplotlib>=3.8.0
pandas>=2.1.0
numpy>=1.24.0
python-binance>=1.0.19
discord-webhook>=1.3.0
python-dotenv>=1.0.0
schedule>=1.2.0
google-generativeai>=0.3.0 